+++ CHANGES.md
@@ -1,6 +1,14 @@
 Changelog for singletons project
 ================================
 
+next
+----
+* Require GHC 8.10.
+* `singletons` now does a more much thorough job of rejecting higher-rank types
+  during promotion or singling, as `singletons` cannot support them.
+  (Previously, `singletons` would sometimes accept them, often changing rank-2
+  types to rank-1 types incorrectly in the process.)
+
 2.6
 ---
 * Require GHC 8.8.